Move static methods to a util class
